RQ-216 restructured API client
