+++ app/src/components/common/APIClient/types.ts
@@ -1,0 +1,8 @@
+import { RequestMethod } from "views/features/apis/types";
+
+export interface APIClientRequest {
+  url: string;
+  headers?: Record<string, string>;
+  method?: RequestMethod;
+  body?: string | FormData;
+}
